+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 ### Changed
 
 - Updated various CONFIG properties and i18n strings to match the final PDF. (#576)
+- [BREAKING] Reworked how kit bonuses are applied to abilities to line up with the new Active Effect type
 - [BREAKING] Redid the implementation of how "Applied Effect" Power Roll Effects work. Now, instead of several sets, you can add an entire effect and it will have multiple properties to configure per tier, such as the Potency Condition, if/how it overrides the end, and any additional properties like "stackable".
 - [BREAKING] Moved recoveries from `system.hero.recoveries` to `system.recoveries`.
 - [BREAKING] Character max recoveries and stamina are now derived from their class item (#627).